requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django>=4.1.7
django-storages==1.13.2
djangorestframework==3.14.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
